+++ src/main/java/roomescape/domain/login/LoginRequest.java
@@ -1,0 +1,5 @@
+package roomescape.domain.login;
+
+public record LoginRequest(String email, String password) {
+
+}
