+++ src/main/java/roomescape/domain/login/LoginResponseDto.java
@@ -1,0 +1,6 @@
+package roomescape.domain.login;
+
+
+public record LoginResponseDto(String name) {
+
+}
