--- src/main/java/roomescape/global/dto/CheckResponseDto.java
@@ -1,6 +1,0 @@
-package roomescape.global.dto;
-
-
-public record CheckResponseDto(String name) {
-
-}
