--- src/main/java/roomescape/global/dto/TokenRequest.java
@@ -1,5 +1,0 @@
-package roomescape.global.dto;
-
-public record TokenRequest(String email, String password) {
-
-}
